Made level 12
Made the 12th level, made the buttons for all the levels in level select.  Made it so all the levels are named on the main node of every level. Made it so level 5 leads to level 6, made a hotkey for restarting (r) and made it an option in the pause menu.